Move prisma and prisma-adapter into lib/prisma
