+++ src/pages/api/auth/[...nextauth].js
@@ -4,7 +4,7 @@ import GitHubProvider from "next-auth/providers/github"
 import GoogleProvider from "next-auth/providers/google";
 
 import { PrismaAdapter } from "@next-auth/prisma-adapter"
-import prisma from "@/lib/prisma"
+import prisma from "@/lib/prisma/prisma"
 
 export const authOptions = {
   adapter: PrismaAdapter(prisma),
